Fix a bug when updating a product
The script was deleting the old thumbnail when submitting an update without uploading a new thumbnail

+++ app/src/Controller/Admin/ProductController.php
@@ -69,8 +69,12 @@ class ProductController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->entityManager->flush(); // needs to save the product's data first
-            $thumbnailService->delete($fileRepository->find($product->getThumbnailId() ?: 0));
-            $this->uploadThumbnail($form, $thumbnailService, $product);
+
+            if ($form->has('thumbnailFile') && $form->get('thumbnailFile')->getData()) {
+                $thumbnailService->delete($fileRepository->find($product->getThumbnailId() ?: 0));
+                $this->uploadThumbnail($form, $thumbnailService, $product);
+            }
+
             $this->addFlash('success', 'Your product has been updated.');
 
             return $this->redirectToRoute('admin_product_index');
